pyproject.toml
[project]
name = "pillow-draw-platform-debug"
version = "0.0.0"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
  "Pillow>=12.0.0" # Image processing library
]
